+++ lib/axios/index.ts
@@ -1,0 +1,3 @@
+export { default as ApiService } from './endpoints'
+export { default as apiClient } from './http'
+export * from './model'
